requirements.txt
academictorrents==2.3.3
albumentations==1.4.10
diffusers==0.21.0.dev0
einops==0.8.0
kornia==0.5.8
lmdeploy==0.4.2
matplotlib==3.9.0
mmengine==0.10.4
natsort==8.4.0
nltk==3.8.1
numpy==2.0.0
omegaconf==2.3.0
opencv_python==4.7.0.72
pandas==2.0.2
Pillow==10.3.0
pycocotools==2.0.7
pytorch_lightning==2.3.1
PyYAML==6.0.1
PyYAML==6.0.1
ram==0.1
Requests==2.32.3
scipy==1.14.0
segment_anything==1.0
skimage==0.0
sng_parser==1.3.0
spacy==3.7.4
submitit==1.5.1
tiktoken==0.6.0
timm==0.4.12
torch==2.3.0
torchvision==0.18.0
tqdm==4.65.2
transformers==4.40.2
transformers_stream_generator==0.0.5
vllm==0.4.2
wandb==0.17.3
webdataset==0.2.86
